Create README.md
The project “Analysis on Medicine and Doctor Availability in Government Hospitals” is designed to check the availability of medicines and doctors in government hospitals. It stores hospital data in a database and helps in tracking shortages, monitoring doctor availability by specialization, and generating useful reports. This system can be used by healthcare administrators to improve hospital resource management and ensure timely availability of medicines and doctors for patients